+++ packages/lumx-react/src/components/thumbnail/Thumbnail.stories.tsx
@@ -1,36 +1,67 @@
 import { Alignment, AspectRatio, CrossOrigin, Size, Theme, Thumbnail, ThumbnailVariant } from '@lumx/react';
+import { htmlDecode } from '@lumx/react/utils/htmlDecode';
 import { boolean, number, select, text } from '@storybook/addon-knobs';
 import React from 'react';
 
 export default { title: 'LumX components/thumbnail/Thumbnail' };
 
-const numberKnobOtions = {
+interface StoryProps {
+    theme: Theme;
+}
+
+const numberKnobOptions = {
     max: 1,
     min: -1,
     range: true,
     step: 0.1,
 };
 
+const corsOptions = {
+    None: undefined,
+    Anonymous: CrossOrigin.anonymous,
+    UseCredentials: CrossOrigin.useCredentials,
+};
+
+const groupId = 'Options';
+
 /**
  * Thumbnail story
  * @return simple Thumbnail.
  */
-export const defaultThumbnail = ({ theme }: { theme: Theme }) => {
+export const defaultThumbnail = ({ theme }: StoryProps) => {
+    const align = select<Alignment>('Alignment', Alignment, Alignment.left, groupId);
+    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, groupId);
+    const isCrossOriginEnabled = boolean('Is CORS enabled', false, groupId);
+    const crossOrigin = select('CORS', corsOptions, corsOptions.None, groupId);
+    const fillHeight = boolean('Fill Height', false, groupId);
+    const focusPoint = {
+        x: number('focusX', 0, numberKnobOptions, groupId),
+        y: number('focusY', 0, numberKnobOptions, groupId),
+    };
+    const image = text(
+        'Url',
+        'https://i.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
+        groupId,
+    );
+    const isFollowingWindowSize = boolean('Update on window resize', true, groupId);
+    const resizeDebounceTime = number('Debounce time after resize', 20, undefined, groupId);
+    const variant = select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, groupId);
+
     return (
         <Thumbnail
-            align={select<Alignment>('Alignment', Alignment, Alignment.left, 'Options')}
-            aspectRatio={select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, 'Options')}
-            isCrossOriginEnabled={boolean('Enable CORS', true, 'Options')}
-            crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
-            fillHeight={boolean('Fill Height', false, 'Options')}
-            focusPoint={{
-                x: number('focusX', 0, numberKnobOtions, 'Options'),
-                y: number('focusY', 0, numberKnobOtions, 'Options'),
-            }}
-            image={text('Url image', 'https://i.picsum.photos/id/1001/2400/1400.jpg', 'Options')}
+            align={align}
+            aspectRatio={aspectRatio}
+            isCrossOriginEnabled={isCrossOriginEnabled}
+            crossOrigin={crossOrigin}
+            fillHeight={fillHeight}
+            focusPoint={focusPoint}
+            image={htmlDecode(image)}
+            isFollowingWindowSize={isFollowingWindowSize}
+            resizeDebounceTime={resizeDebounceTime}
             size={select(
                 'Size',
                 {
+                    None: undefined,
                     XXS: Size.xxs,
                     XS: Size.xs,
                     S: Size.s,
@@ -40,37 +71,50 @@ export const defaultThumbnail = ({ theme }: { theme: Theme }) => {
                     XXL: Size.xxl,
                 },
                 Size.xxl,
-                'Options',
+                groupId,
             )}
             theme={theme}
-            isFollowingWindowSize={boolean('Update on window resize', true, 'Options')}
-            resizeDebounceTime={number('Debounce time after resize', 20, undefined, 'Options')}
-            variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
+            variant={variant}
         />
     );
 };
 
-export const thumbnailWithCustomImgProps = ({ theme }: { theme: Theme }) => {
+export const thumbnailWithCustomImgProps = ({ theme }: StoryProps) => {
     const [image, setImage] = React.useState('https://not-found');
     const onError = () => {
-        setImage('https://i.picsum.photos/id/1001/2400/1400.jpg');
+        // tslint:disable-next-line: no-console
+        console.log('onError event called');
+        setImage('https://i.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0');
     };
+
+    const align = select<Alignment>('Alignment', Alignment, Alignment.left, groupId);
+    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, groupId);
+    const isCrossOriginEnabled = boolean('Is CORS enabled', false, groupId);
+    const crossOrigin = select('CORS', corsOptions, corsOptions.Anonymous, groupId);
+    const fillHeight = boolean('Fill Height', false, groupId);
+    const focusPoint = {
+        x: number('focusX', 0, numberKnobOptions, groupId),
+        y: number('focusY', 0, numberKnobOptions, groupId),
+    };
+    const isFollowingWindowSize = boolean('Update on window resize', true, groupId);
+    const resizeDebounceTime = number('Debounce time after resize', 20, undefined, groupId);
+    const variant = select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, groupId);
 
     return (
         <Thumbnail
-            align={select<Alignment>('Alignment', Alignment, Alignment.left, 'Options')}
-            aspectRatio={select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, 'Options')}
-            isCrossOriginEnabled={boolean('Enable CORS', true, 'Options')}
-            crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
-            fillHeight={boolean('Fill Height', false, 'Options')}
-            focusPoint={{
-                x: number('focusX', 0, numberKnobOtions, 'Options'),
-                y: number('focusY', 0, numberKnobOtions, 'Options'),
-            }}
+            align={align}
+            aspectRatio={aspectRatio}
+            isCrossOriginEnabled={isCrossOriginEnabled}
+            crossOrigin={crossOrigin}
+            fillHeight={fillHeight}
+            focusPoint={focusPoint}
             image={image}
+            imgProps={{ onError }}
+            isFollowingWindowSize={isFollowingWindowSize}
             size={select(
                 'Size',
                 {
+                    None: undefined,
                     XXS: Size.xxs,
                     XS: Size.xs,
                     S: Size.s,
@@ -80,26 +124,38 @@ export const thumbnailWithCustomImgProps = ({ theme }: { theme: Theme }) => {
                     XXL: Size.xxl,
                 },
                 Size.xxl,
-                'Options',
+                groupId,
             )}
+            resizeDebounceTime={resizeDebounceTime}
             theme={theme}
-            isFollowingWindowSize={boolean('Update on window resize', true, 'Options')}
-            resizeDebounceTime={number('Debounce time after resize', 20, undefined, 'Options')}
-            variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
-            imgProps={{ onError }}
+            variant={variant}
         />
     );
 };
 
-export const brokenThumbnailWithFallbackProps = ({ theme }: { theme: Theme }) => (
-    <Thumbnail
-        align={select<Alignment>('Alignment', Alignment, Alignment.left, 'Options')}
-        aspectRatio={select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, 'Options')}
-        isCrossOriginEnabled={boolean('Enable CORS', true, 'Options')}
-        crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
-        fillHeight={boolean('Fill Height', false, 'Options')}
-        image={text('Url image', 'https://i.picsum.photos/id/1001/2400/1400.jp', 'Options')}
-        theme={theme}
-        variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
-    />
-);
+export const brokenThumbnailWithFallbackProps = ({ theme }: StoryProps) => {
+    const align = select<Alignment>('Alignment', Alignment, Alignment.left, groupId);
+    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, groupId);
+    const isCrossOriginEnabled = boolean('Is CORS enabled', false, groupId);
+    const crossOrigin = select('CORS', corsOptions, corsOptions.None, groupId);
+    const fillHeight = boolean('Fill Height', false, groupId);
+    const image = text(
+        'Url',
+        'https://i.picsum.photos/id/237/536/354.jp?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
+        groupId,
+    );
+    const variant = select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, groupId);
+
+    return (
+        <Thumbnail
+            align={align}
+            aspectRatio={aspectRatio}
+            isCrossOriginEnabled={isCrossOriginEnabled}
+            crossOrigin={crossOrigin}
+            fillHeight={fillHeight}
+            image={htmlDecode(image)}
+            theme={theme}
+            variant={variant}
+        />
+    );
+};